Fix: search functionality complete along with several other bugfixes
Search Functionality now displays books that do not have an imageLink or author. Search no longer breaks main page after being used and returning. Proptypes included for all components now. Select option on search page fixed so books that are not shelved are marked as none in the dropdown. This should be the final commit that includes all the necessary project features.